requirements.txt
allure-pytest==2.11.0
allure-python-commons==2.11.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.4
iniconfig==1.1.1
loguru==0.6.0
packaging==21.3
playwright==1.27.0
pluggy==1.0.0
py==1.11.0
pyee==8.1.0
pyparsing==3.0.9
pytest==7.1.3
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dotenv==0.21.0
python-slugify==6.1.2
requests==2.28.1
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.12
